Project 4
Largest Non-consecutive Subset Of Array....
You’ll be given an array of (positive and negative) integers with size N. Your goal is to find the
subset of the entries with the largest sum with 1 important restriction: no 2 elements of the subset can be
consecutive. You will find an O(N) time dynamic programming algorithm.